Refactor: Separate API services and utility modules
- Moved API client scripts (Bithumb, Coinone, Korbit, Upbit, Solana) into a new `services` directory.
- Renamed API client files for consistency (e.g., `api_bithumb.py` to `services/bithumb_api.py`).
- Moved `api_prices.py` to `utils/price_fetcher.py`.
- Created `__init__.py` in `services` and `utils` directories to make them packages.
- Updated all import statements in `agg.py` and affected modules to reflect the new structure.
